Allow passing hash with URL string and :headers
Currently, if you need to pass the :headers hash, you must pass the URL
as a hash; this means if you have a URL that's a string, you have to
break it apart so that CurbFu can put it back together again.

The API here isn't ideal -- it's not clear to me why headers are passed
in what seems to be the URL hash. but, this works, and is compatible
with the current API.